Pull up AbstractHerokuTask
