+++ src/main/java/com/heroku/AbstractDeploymentTask.java
@@ -4,18 +4,13 @@ import com.atlassian.bamboo.build.logger.BuildLogger;
 import com.atlassian.bamboo.security.StringEncrypter;
 import com.atlassian.bamboo.task.*;
 import com.heroku.api.App;
-import com.heroku.api.Heroku;
 import com.heroku.api.HerokuAPI;
-import com.heroku.api.exception.RequestFailedException;
 import com.herokuapp.directto.client.DeployRequest;
 import com.herokuapp.directto.client.DirectToHerokuClient;
 import com.herokuapp.directto.client.EventSubscription;
 import com.herokuapp.directto.client.VerificationException;
-import org.apache.tools.ant.BuildListener;
-import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
-import java.net.HttpURLConnection;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -27,55 +22,24 @@ import static com.herokuapp.directto.client.EventSubscription.Subscriber;
 /**
  * @author Ryan Brainard
  */
-public abstract class AbstractDeploymentTask<P extends DeploymentPipeline> implements TaskType {
+public abstract class AbstractDeploymentTask<P extends DeploymentPipeline> extends AbstractHerokuTask {
 
-    /**
-     * A sandbox for static methods that don't play well with jMock
-     */
-    protected static interface StaticSandbox {
-        TaskResult success(TaskContext taskContext);
-        TaskResult failed(TaskContext taskContext);
-    }
-    
     private final P pipeline;
-    private final StaticSandbox staticSandbox;
 
     public AbstractDeploymentTask(P pipeline) {
-        this(pipeline, new StaticSandbox() {
-            @Override
-            public TaskResult success(TaskContext taskContext) {
-                return TaskResultBuilder.create(taskContext).success().build();
-            }
-
-            @Override
-            public TaskResult failed(TaskContext taskContext) {
-                return TaskResultBuilder.create(taskContext).failed().build();
-            }
-        });
+        super();
+        this.pipeline = pipeline;
     }
 
     public AbstractDeploymentTask(P pipeline, StaticSandbox staticSandbox) {
+        super(staticSandbox);
         this.pipeline = pipeline;
-        this.staticSandbox = staticSandbox;
     }
 
-    @NotNull
     @Override
-    public TaskResult execute(@NotNull final TaskContext taskContext) throws TaskException {
-        try {
-            return executeInternal(taskContext);
-        } catch (HerokuBambooHandledException e) {
-            taskContext.getBuildLogger().addErrorLogEntry(e.getMessage());
-            return staticSandbox.failed(taskContext);
-        }
-    }
-
-    private TaskResult executeInternal(TaskContext taskContext) {
+    protected TaskResult execute(TaskContext taskContext, String apiKey, HerokuAPI api, App app) {
         final BuildLogger buildLogger = taskContext.getBuildLogger();
-        final String apiKey = new StringEncrypter().decrypt(taskContext.getConfigurationMap().get(AbstractDeploymentTaskConfigurator.API_KEY));
         final String pipelineName = pipeline.getPipelineName();
-        final HerokuAPI api = new HerokuAPI(apiKey);
-        final App app = getOrCreateApp(buildLogger, api, taskContext.getConfigurationMap().get("appName"));
         final DirectToHerokuClient client = new DirectToHerokuClient.Builder()
                 .setApiKey(apiKey)
                 .setConsumersUserAgent(HerokuPluginProperties.getUserAgent())
@@ -121,36 +85,6 @@ public abstract class AbstractDeploymentTask<P extends DeploymentPipeline> imple
         return "success".equals(deployResults.get("status"))
                 ? staticSandbox.success(taskContext)
                 : staticSandbox.failed(taskContext);
-    }
-
-    protected App getOrCreateApp(BuildLogger buildLogger, HerokuAPI api, String appName) {
-        App app;
-
-        try {
-            app = api.getApp(appName);
-        } catch (RequestFailedException appListingException) {
-            if (appListingException.getStatusCode() == HttpURLConnection.HTTP_FORBIDDEN) {
-                throw new HerokuBambooHandledException("No access to Heroku app '" + appName + "'. Check API key, app name, and ensure you have access.");
-            }
-
-            try {
-                app = api.createApp(new App().named(appName).on(Heroku.Stack.Cedar));
-                buildLogger.addBuildLogEntry("Created new app " + appName);
-            } catch (RequestFailedException appCreationException) {
-                if (appCreationException.getStatusCode() == HttpURLConnection.HTTP_UNAUTHORIZED) {
-                    throw new HerokuBambooHandledException("No access to create Heroku app '" + appName + "'. Check API key.");
-                }
-
-                buildLogger.addErrorLogEntry("Unknown error creating app '" + appName + "'\n" + appCreationException.getMessage());
-                throw appCreationException;
-            }
-        }
-
-        if (app == null || app.getId() == null) {
-            throw new HerokuBambooHandledException("Heroku app '" + appName + "' could not be found. Check API key, app name, and ensure you have access.");
-        }
-
-        return app;
     }
 
 }
