+++ src/main/java/com/heroku/AnvilPush.java
@@ -1,0 +1,19 @@
+package com.heroku;
+
+import com.atlassian.bamboo.task.TaskContext;
+import com.atlassian.bamboo.task.TaskException;
+import com.atlassian.bamboo.task.TaskResult;
+import com.atlassian.bamboo.task.TaskType;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * @author Ryan Brainard
+ */
+public class AnvilPush implements TaskType {
+
+    @NotNull
+    @Override
+    public TaskResult execute(@NotNull TaskContext taskContext) throws TaskException {
+        return null;  //To change body of implemented methods use File | Settings | File Templates.
+    }
+}
